Refactor worker APIs into more granular namespaces.
